refactor(app): extract UI/cursor loading and level load/save modules
Move cursor/help/autocomplete loading, file dialog, custom cursor, and progress overlay (with LoadOneSpr) into app_ui.cpp; and level load/save, game-level sync, ATTA flush, texture-map export (LoadLevel, SetGameLevel, FlushAttaProxiesToMef, SaveCurrentLevel, ExportTextureMap) into app_level.cpp. app.cpp core 1376 -> 719 lines (now mainly ctor/Init/Shutdown, window subclass, draw-state toggles, OnIdle/Frame). Pure code motion: builds clean, runtime-verified (level 5, 910 objects, no errors).
