+++ source/app_level.cpp
@@ -1,0 +1,314 @@
+/******************************************************************************
+ * @file    app_level.cpp
+ * @brief   App: level load/save, game-level sync, ATTA flush, texture-map export
+ *          Split from app.cpp; shares app_internal.h.
+ *****************************************************************************/
+#include "app_internal.h"
+
+void App::LoadLevel(int level_no) {
+	try {
+		Logger::Get().Log(LogLevel::INFO, "[App] ==========================================");
+		Logger::Get().Log(LogLevel::INFO, "[App] LoadLevel() START for level " + std::to_string(level_no));
+		Logger::Get().Log(LogLevel::INFO, "[App] ==========================================");
+
+		// Track in-session level switches. The FIRST load (last_loaded_level_ < 0) is
+		// equivalent to a fresh process; any later load is a MENU/RELOAD switch whose
+		// only difference from a fresh process is leftover per-level state.
+		const bool is_switch = (last_loaded_level_ >= 0 && last_loaded_level_ != level_no);
+		if (is_switch) {
+			Logger::Get().Log(LogLevel::INFO, "[App] MENU/RELOAD switch from level " +
+				std::to_string(last_loaded_level_) + " to " + std::to_string(level_no) +
+				" — performing full previous-level teardown");
+			AssetExtractor::ClearLevelAssets(last_loaded_level_, Utils::GetExeDirectory());
+		} else {
+			Logger::Get().Log(LogLevel::INFO, "[App] Initial level load (level " +
+				std::to_string(level_no) + ")");
+		}
+
+		AssetExtractor::ClearLevelAssets(level_no, Utils::GetExeDirectory());
+
+		// Verify level number is valid
+		if (level_no < MIN_LEVEL_NO || level_no > MAX_LEVEL_NO) {
+			std::string errorMsg = "Invalid level number: " + std::to_string(level_no) + " (valid range: " + std::to_string(MIN_LEVEL_NO) + "-" + std::to_string(MAX_LEVEL_NO) + ")";
+			Utils::LogAndShowError(errorMsg, "IGI Editor - Error");
+			Logger::Get().Log(LogLevel::ERR, errorMsg);
+			// Exit the application safely; destructor will be called automatically
+			std::exit(EXIT_FAILURE);
+		}
+		if (Config::Get().enableBackup) {
+			std::string gameLevelDir = Utils::GetIGIRootPath() + "\\missions\\location0\\level" + std::to_string(level_no);
+			std::string backupLevelDir = Utils::GetExeDirectory() + "\\editor\\backup\\level" + std::to_string(level_no);
+			if (!std::filesystem::exists(backupLevelDir) && std::filesystem::exists(gameLevelDir)) {
+				try {
+					std::filesystem::create_directories(backupLevelDir);
+					std::filesystem::copy(gameLevelDir, backupLevelDir, std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+					Logger::Get().Log(LogLevel::INFO, "[App] Backup created for level " + std::to_string(level_no) + " at " + backupLevelDir);
+				} catch (const std::exception& e) {
+					Logger::Get().Log(LogLevel::ERR, "[App] Failed to create level backup: " + std::string(e.what()));
+				}
+			}
+		}
+		
+		selected_object_index_ = -1;
+		hover_object_index_ = -1;
+		status_message_.clear();
+
+		// NOTE: We must NOT purge the previous level's extracted assets here. The
+		// texture/model resolver (FindTextureFile step 6 + lazy cross-level extraction)
+		// deliberately searches *other* levels' extracted folders to resolve cross-level
+		// references (e.g. a level-6 object using a level-14 texture). Deleting them on
+		// switch removes legitimate fallback sources. The renderer caches are still fully
+		// torn down by BeginLoadLevel()/ClearCaches() below.
+		renderer_.SetLevel(level_no);
+		renderer_.BeginLoadLevel();
+		Logger::Get().Log(LogLevel::INFO, "[App] After BeginLoadLevel teardown: renderer meshCache=" +
+			std::to_string(renderer_.GetMeshCacheCount()) + " textureCache=" +
+			std::to_string(renderer_.GetTextureCacheCount()) + " (both should be 0)");
+		renderer_.SetSplineTerrainQuery([this](double x, double y, float& z) {
+			return level_.GetTerrainZ(x, y, z);
+		});
+
+		Level::load_params_s level_load_params_s = {
+			.level_no_ = level_no,
+			.render_res_loader_ = &renderer_
+		};
+
+		// ── Loading overlay ──────────────────────────────────────────────────
+		// Staged progress: the load is synchronous, so we redraw + swap the bar at
+		// each milestone (10% → 100%). Reusable lambda fills proportionally to pct.
+		auto drawLoadBar = [&](int pct, const char* stage) {
+			char title[32];
+			snprintf(title, sizeof(title), "Loading Level %d", level_no);
+			DrawProgressOverlay(title, pct, stage);
+		};
+		drawLoadBar(10, "reading level");
+		// ─────────────────────────────────────────────────────────────────────
+
+		drawLoadBar(25, "models & terrain");
+		glm::vec3 start_pos;
+		float start_yaw;
+		if (level_.Load(level_load_params_s, start_pos, start_yaw)) {
+			const auto& config = Config::Get();
+			viewer_.pos_ = (config.cameraPosX != 0.0f || config.cameraPosY != 0.0f || config.cameraPosZ != 0.0f) ?
+				glm::vec3(config.cameraPosX, config.cameraPosY, config.cameraPosZ) : start_pos;
+
+			bool hasConfigOri = (config.cameraOriX != 0.0f || config.cameraOriY != 0.0f || config.cameraOriZ != 0.0f);
+			if (hasConfigOri) {
+				viewer_.yaw_   = config.cameraOriX;
+				viewer_.pitch_ = config.cameraOriY;
+				viewer_.roll_  = config.cameraOriZ;
+			} else {
+				// Convert game yaw (radians) to viewer yaw (degrees, 0=+Y north, CW).
+				// Add 180° so the editor camera faces toward objects rather than with them.
+				viewer_.yaw_   = glm::degrees(atan2f(-cosf(start_yaw), sinf(start_yaw))) + 180.0f;
+				viewer_.pitch_ = 10.0f;
+				viewer_.roll_  = 0.0f;
+			}
+
+			UpdateViewerVectors();
+			Logger::Get().Log(LogLevel::INFO, "[App] Level " + std::to_string(level_no) + " loaded. Viewer start=(" + std::to_string(viewer_.pos_.x) + "," + std::to_string(viewer_.pos_.y) + "," + std::to_string(viewer_.pos_.z) + ") yaw=" + std::to_string(viewer_.yaw_));
+			last_loaded_level_ = level_no;
+		}
+		else {
+			std::string errorMsg = "Failed to load level " + std::to_string(level_no) + "\n\nPlease check if the terrain files exist in the correct location.";
+			Utils::ShowError(errorMsg, "IGI Editor - Error");
+			Logger::Get().Log(LogLevel::ERR, "[App] Failed to load level " + std::to_string(level_no));
+		}
+
+
+		drawLoadBar(60, "tracks");
+		// Step 2: Track Evaluation (Dynamic object placement along paths)
+		EvaluateTrainTrackPositions();
+
+		drawLoadBar(75, "snapping objects");
+		// Step 3: Always snap objects to terrain after any level load
+		Logger::Get().Log(LogLevel::INFO, "[App] Step 3: Snapping objects to terrain...");
+		SnapObjectsToTerrain();
+		drawLoadBar(90, "finalizing");
+
+		// AI rotation override: AI models (HumanSoldier, HumanAI) only have horizontal rotation
+		auto& objects = level_.GetLevelObjects().GetObjects();
+
+		// Build runtime parameter schemas from this level's Task_DeclareParameters so
+		// the property editor can expose EVERY declared field of any task type.
+		TaskSchemaNS::ClearRegisteredSchemas();
+		for (const auto& obj : objects) {
+			if (obj.type == "Task_DeclareParameters" && obj.argTokens.size() >= 3) {
+				std::string typeName = obj.argTokens[0];
+				if (typeName.size() >= 2 && typeName.front() == '"' && typeName.back() == '"')
+					typeName = typeName.substr(1, typeName.size() - 2);
+				TaskSchemaNS::RegisterSchema(typeName, TaskSchemaNS::ParseDeclaration(obj.argTokens));
+			}
+		}
+
+		for (auto& obj : objects) {
+			if (obj.modelId == "000_01_1") continue; // Skip Player Jones
+			if (obj.type == "HumanSoldier" || obj.type == "HumanAI" || obj.type.find("AITYPE") == 0) {
+				obj.rot.x = 0.0;           // PITCH = 0
+				obj.rot.y = 0.0;           // ROLL = 0
+				// Preserve existing rotation if it's already set, otherwise default to a full circle
+				if (obj.rot.z == 0.0) obj.rot.z = 6.28318;
+				Logger::Get().Log(LogLevel::INFO, "[App] Applied AI rotation override (horizontal only) for " + obj.name + " (" + obj.type + ")");
+			}
+		}
+
+		// Log all loaded objects for verification script
+		for (const auto& obj : objects) {
+			if (obj.isSplineWaypoint || !obj.segmentModelId.empty()) {
+			    Logger::Get().Log(LogLevel::INFO, "[App_Debug] Found waypoint/segment. modelId=" + obj.modelId + " segmentModelId=" + obj.segmentModelId);
+			}
+			std::string mId = !obj.modelId.empty() ? obj.modelId : obj.segmentModelId;
+			if (obj.deleted || mId.empty()) continue;
+			Logger::Get().Log(LogLevel::INFO, "[LevelLoader] Object Loaded: ModelID=" + mId +
+				", Type=" + obj.type + ", Name=" + obj.name + ", Pos=(" +
+				std::to_string(obj.pos.x) + ", " + std::to_string(obj.pos.y) + ", " + std::to_string(obj.pos.z) + ")" +
+				", Ori=(" + std::to_string(obj.rot.x) + ", " + std::to_string(obj.rot.y) + ", " + std::to_string(obj.rot.z) + ")" +
+				", Tex=" + mId + ", Model=" + mId);		}
+		
+		RebuildLevelModelIds();
+
+		// Build the set of models packed in this level's .res (names only — stream so we
+		// never buffer the whole 200+MB archive in the 32-bit process). (issue 2)
+		{
+			std::string gameRes = Utils::GetIGIRootPath() +
+				"\\missions\\location0\\level" + std::to_string(level_no) +
+				"\\models\\level" + std::to_string(level_no) + ".res";
+			level_res_models_ = ResModelSet();
+			std::string resErr;
+			size_t entryCount = 0;
+			bool ok = RES_ForEachEntry(gameRes,
+				[&](const std::string& name, const uint8_t*, size_t) {
+					level_res_models_.AddEntry(name);
+					++entryCount;
+				}, resErr);
+			Logger::Get().Log(LogLevel::INFO, std::string("[App] Level .res model set: ") +
+				(ok ? std::to_string(entryCount) + " entries (streamed)"
+				    : "UNAVAILABLE (" + gameRes + "): " + resErr));
+		}
+		Logger::Get().Log(LogLevel::INFO, "[App] ==========================================");
+		Logger::Get().Log(LogLevel::INFO, "[App] LoadLevel() COMPLETE for level " + std::to_string(level_no));
+		Logger::Get().Log(LogLevel::INFO, "[App] ==========================================");
+		drawLoadBar(100, "done");
+	}
+	catch (const std::exception& e) {
+		std::string errorMsg = "Error loading level " + std::to_string(level_no) + ":\n" + std::string(e.what());
+		Utils::LogAndShowError(errorMsg, "IGI Editor - Error");
+		Logger::Get().Log(LogLevel::ERR, errorMsg);
+	}
+	catch (...) {
+		std::string errorMsg = "Unknown error loading level " + std::to_string(level_no);
+		Utils::LogAndShowError(errorMsg, "IGI Editor - Error");
+		Logger::Get().Log(LogLevel::ERR, errorMsg);
+	}
+}
+
+void App::SetGameLevel(int level_no) {
+	bridge_.SetGameLevel(level_no);
+}
+
+
+
+
+
+
+
+
+void App::FlushAttaProxiesToMef() {
+	for (auto& obj : level_.GetLevelObjects().GetObjects()) {
+		if (!obj.isAttaProxy || !obj.modified) continue;
+
+		// Position: convert world pos → local pos via stored inverse parent matrix.
+		glm::vec4 lp = obj.attaInvParentMat * glm::vec4(glm::vec3(obj.pos), 1.0f);
+		glm::vec3 localPos(lp);
+
+		// Rotation: rebuild world rotation matrix from Euler angles (Rz * Rx * Ry),
+		// then convert to local rotation: localRot = parentRotInv * worldRot.
+		// The upper-left 3×3 of attaInvParentMat IS parentRotInv.
+		glm::mat4 wr(1.0f);
+		wr = glm::rotate(wr, (float)obj.rot.z, glm::vec3(0,0,1));
+		wr = glm::rotate(wr, (float)obj.rot.x, glm::vec3(1,0,0));
+		wr = glm::rotate(wr, (float)obj.rot.y, glm::vec3(0,1,0));
+		glm::mat3 worldRot3(wr);
+		glm::mat3 parentRotInv = glm::mat3(obj.attaInvParentMat);
+		glm::mat3 localRot = parentRotInv * worldRot3;
+
+		renderer_.UpdateAttaLocalPosInMef(obj.attaParentModelId, obj.attaIsBuilding,
+		                                  obj.attaRecordIndex, localPos, localRot);
+		obj.modified = false;
+	}
+}
+
+void App::SaveCurrentLevel() {
+	try {
+		Logger::Get().Log(LogLevel::INFO, "[App] SaveCurrentLevel() called");
+		FlushAttaProxiesToMef();
+		level_.SaveChanges();
+
+		// Compile edited AI script (.qsc text -> .qvm file) before saving the level QVM.
+		if (ai_script_dirty_ && !ai_script_path_.empty()) {
+			Logger::Get().Log(LogLevel::INFO, "[App] Compiling modified AI script to: " + ai_script_path_);
+			auto lexResult   = qsc::Lex(ai_script_text_);
+			auto parseResult = lexResult.ok ? qsc::Parse(lexResult.tokens) : qsc::ParseResult{};
+			std::string compileErr;
+			bool ok = lexResult.ok && parseResult.ok &&
+			          qvm::CompileToFile(*parseResult.program, ai_script_path_, &compileErr);
+			if (ok) {
+				QVMFile check = QVM_Parse(ai_script_path_);
+				if (check.valid) {
+					ai_script_dirty_ = false;
+					status_message_ = "AI script compiled: " + ai_script_path_;
+					Logger::Get().Log(LogLevel::INFO, "[App] AI script compiled OK");
+				} else {
+					Logger::Get().Log(LogLevel::ERR, "[App] AI script round-trip failed -- file may be corrupt");
+					status_message_ = "AI script compile: round-trip failed";
+				}
+			} else {
+				std::string detail = compileErr.empty() ? "(no detail)" : compileErr;
+				Logger::Get().Log(LogLevel::ERR, "[App] AI script compile error: " + detail);
+				status_message_ = "AI script compile error: " + detail;
+			}
+		}
+
+		Logger::Get().Log(LogLevel::INFO, "[App] Calling SaveAndCompile()");
+		SaveAndCompile();
+	}
+	catch (const std::exception& e) {
+		std::string errorMsg = "Error saving level:\n" + std::string(e.what());
+		Utils::LogAndShowError(errorMsg, "IGI Editor - Error");
+		Logger::Get().Log(LogLevel::ERR, errorMsg);
+	}
+	catch (...) {
+		std::string errorMsg = "Unknown error saving level";
+		Utils::LogAndShowError(errorMsg, "IGI Editor - Error");
+		Logger::Get().Log(LogLevel::ERR, errorMsg);
+	}
+}
+
+void App::ExportTextureMap() {
+	int levelNo = level_.GetLevelNo();
+	const std::string root = Utils::GetIGIRootPath();
+	const std::string datPath = root + "\\missions\\location0\\level" +
+	    std::to_string(levelNo) + "\\level" + std::to_string(levelNo) + ".dat";
+	const std::string outPath = Utils::GetExeDirectory() +
+	    "\\level" + std::to_string(levelNo) + "_texmap.json";
+
+	Logger::Get().Log(LogLevel::INFO,
+	    "[App] ExportTextureMap level=" + std::to_string(levelNo) +
+	    " dat=" + datPath + " out=" + outPath);
+
+	DATFile dat = DAT_Parse(datPath);
+	if (!dat.valid) {
+		Utils::ShowError("Failed to parse DAT:\n" + dat.error, "Export Texture Map");
+		return;
+	}
+	if (DAT_WriteJSON(dat, outPath)) {
+		Utils::ShowInfo(
+		    "Texture map exported (JSON):\n" + outPath +
+		    "\nModels: " + std::to_string(dat.models.size()) +
+		    "  Textures: " + std::to_string(dat.allTextures.size()),
+		    "Export Texture Map");
+	} else {
+		Utils::ShowError("Could not write to:\n" + outPath, "Export Texture Map");
+	}
+}
+
